+++ src/libIntegrate/_2D/RiemannRule.hpp
@@ -36,10 +36,10 @@ T RiemannRule<T>::operator()( F f, X xa, X xb, size_t xN, X ya, X yb, size_t yN 
   X dx = (xb-xa)/xN; // make sure we don't get integer rounding
   X dy = (yb-ya)/yN; // make sure we don't get integer rounding
   X y = ya;
-  for(int i = 0; i < xN; i++)
+  for(size_t i = 0; i < xN; i++)
   {
     X x = xa;
-    for(int j = 0; j < yN; j++)
+    for(size_t j = 0; j < yN; j++)
     {
       sum += f(x,y);
       x += dx;
@@ -55,7 +55,7 @@ template<typename X, typename Y, typename F>
 T RiemannRule<T>::operator()( X &x, Y &y, F &f ) const
 {
   T sum = 0;
-  for(int i = 0; i < x.size()-1; i++)
+  for(size_t i = 0; i < x.size()-1; i++)
     for(int j = 0; j < y.size()-1; j++)
       sum += f[i][j]*(x[i+1]-x[i])*(y[j+1]-y[j]);
